+++ zenml/components/split_gen/component.py
@@ -21,9 +21,9 @@ from tfx.types import standard_artifacts
 from tfx.types.component_spec import ComponentSpec, ExecutionParameter, \
     ChannelParameter
 
-from zenml.core.components.split_gen import constants
-from zenml.core.components.split_gen import executor
-from zenml.core.standards.standard_keys import StepKeys
+from zenml.components.split_gen import constants
+from zenml.components.split_gen import executor
+from zenml.standards.standard_keys import StepKeys
 
 
 class SplitGenComponentSpec(ComponentSpec):
