+++ zenml/components/tokenizer/component.py
@@ -20,10 +20,10 @@ from tfx.types import standard_artifacts, Channel
 from tfx.types.component_spec import ComponentSpec, ExecutionParameter, \
     ChannelParameter
 
-from zenml.core.components.tokenizer.constants import EXAMPLES, TOKENIZER, \
+from zenml.components.tokenizer.constants import EXAMPLES, TOKENIZER, \
     OUTPUT_EXAMPLES
-from zenml.core.components.tokenizer.executor import TokenizerExecutor
-from zenml.core.standards.standard_keys import StepKeys
+from zenml.components.tokenizer.executor import TokenizerExecutor
+from zenml.standards.standard_keys import StepKeys
 
 
 class TokenizerSpec(ComponentSpec):
