+++ zenml/components/tokenizer/executor.py
@@ -23,10 +23,10 @@ from tfx.types import artifact_utils
 from tfx.types.artifact_utils import get_split_uri
 from tfx.utils import io_utils
 
-from zenml.core.components.split_gen.executor import WriteSplit
-from zenml.core.standards.standard_keys import StepKeys
-from zenml.core.steps.split.utils import get_categorical_value
-from zenml.core.steps.tokenizer import BaseTokenizer
+from zenml.components.split_gen.executor import WriteSplit
+from zenml.standards.standard_keys import StepKeys
+from zenml.steps.split.utils import get_categorical_value
+from zenml.steps.tokenizer import BaseTokenizer
 from zenml.utils import path_utils, source_utils
